Buildable with aarch64-linux-*-gcc
Modified the build script to deal with gcc that defaults to -fPIC (generating
a global offset table in the elf file).

Reorganised how memset is brought into the compiliation.

Removed an assumption about how the address of _start would be calculated.
This might be better done differently, but the EL3 and EL1 kernels run at
different addresses. PC-relative is still the way to go.

Still doesn't run without a modified qemu that loads the "bios" at 0 and
allocates fb memory on a 2MiB boundary.

Changed the build when generating for use with qemu; afaics it still doesn't
allow for an image file to be treated as an SD card at the hardware level.

+++ build.sh
@@ -1,7 +1,13 @@
 #! /bin/bash -
 
+if [ "x$QEMU" != "x" ]; then
+  pushd unit_tests/virtualisation &&
+  ./make_test_image > ../../include/arm32_code.h &&
+  popd || exit 1
+fi
+
 if [ ! -e tools/interfaces -o tools/interfaces.c -nt tools/interfaces ] ; then
-  gcc tools/interfaces.c -o tools/interfaces && ln -sf interfaces tools/server && ln -sf interfaces tools/client
+  gcc tools/interfaces.c -o tools/interfaces && ln -sf interfaces tools/server && ln -sf interfaces tools/client || exit 1
 fi
 
 rm -f kernel8.dump kernel8.elf sdfat/kernel8.img built_drivers/* include/interfaces/client/* include/interfaces/provider/*
@@ -17,24 +23,49 @@ rm -f kernel8.dump kernel8.elf sdfat/kernel8.img built_drivers/* include/interfa
   cd ..
 )
 
-TARGET=aarch64-none-elf-
+TARGET=aarch64-linux-gnu-
+GCC=aarch64-linux-gnu-gcc-10
 
-OPTIMISATION="-Os"
+# Later versions of gcc insert calls to memset when optimising
+OPTIMISATION="-Os memset.c"
 
 # Automatically allow for modifications to Core structure
 core_size() {
-( echo '#include "include/kernel.h"'; echo 'Core var = { 0 };' ) | ${TARGET}gcc -o /tmp/$$.o -x c - -c -DCORE_STACK_SIZE=$1 &&
-${TARGET}objdump --disassemble-all /tmp/$$.o -x | sed -n 's/^.*bss[^0]*0*\(.*\) var$/0x\1/p' || ( rm /tmp/$$.o ; exit 1 )
-rm /tmp/$$.o
+( echo '#include "include/kernel.h"'; echo 'Core core_var = { 0 };' ) | ${GCC} -o /tmp/$$.o -x c - -c -DCORE_STACK_SIZE=$1 -fno-zero-initialized-in-bss &&
+${TARGET}objdump --disassemble-all /tmp/$$.o -x | sed -n 's/^.*data[^0]*0*\(.*\) core_var$/0x\1/p'
 }
 
-# -mgeneral-regs-only Stops the compiler using floating point registers as temprary storage
-# -ffixed-x18 stops the compiler from using x18, so that it can be used to store the current thread (not trusted by the kernel, of course)
-# I anticipate code with lots of fast locks, so a register is probably a better choice than TLS. ICBW.
-CFLAGS="-I include -mgeneral-regs-only $OPTIMISATION -g -Wall -Wextra -fno-zero-initialized-in-bss -nostartfiles -nostdlib -mtune=cortex-a53 -ffixed-x18 -Wno-unused-function $QEMU "
+CFLAGS="-I include $OPTIMISATION"
 
-KERNEL_ELEMENTS="boot.c el3.c el3_gpio4_debug.c secure_el1.c kernel_translation_tables.c memset.c"
-KERNEL_ELEMENTS="boot.c el3.c el3_virtual_machines.c secure_el1.c kernel_translation_tables.c memset.c"
+# Bare metal
+CFLAGS+=" -nostartfiles -nostdlib -nostdinc -static "
+
+# Just the right amount of warnings
+CFLAGS+=" -Wall -Wextra -Wno-unused-function "
+
+# No floating point registers in kernel code, and reserve x18 for task ID
+# (not trusted by the kernel, of course)
+# I anticipate code with lots of fast locks, so a register is probably a
+# better choice than TLS. ICBW.
+CFLAGS+=" -mgeneral-regs-only -ffixed-x18 "
+
+# No loader needed, please. (No .got, .got.plt, .bss sections)
+CFLAGS+=" -fno-zero-initialized-in-bss -fPIE -fpie "
+
+# Where are we going to run it?
+CFLAGS+=" -mtune=cortex-a53 $QEMU "
+
+# Linker flags
+CFLAGS+=" -Wl,--build-id=none,-no-dynamic-linker"
+
+# We do not want a GLOBAL_OFFSET_TABLE (if using a aarch64-linux-...-gcc, the
+# default is -fPIC)
+CFLAGS+=" -fno-PIC"
+
+echo CFLAGS: $CFLAGS
+
+#KERNEL_ELEMENTS="boot.c el3.c el3_gpio4_debug.c secure_el1.c kernel_translation_tables.c"
+KERNEL_ELEMENTS="boot.c el3.c el3_virtual_machines.c secure_el1.c kernel_translation_tables.c"
 
 SYSTEM_DRIVER=system
 MEMORY_DRIVER=physical_memory_allocator
@@ -52,6 +83,7 @@ symbol() {
 }
 
 build_driver() {
+  echo Building $2
   # Each driver's code and data is padded to a 4k boundary, the .bin file will be a multiple of 4k in size.
   # Parameters: ld.script for build (used to locate code other than at 0), driver name
 
@@ -63,9 +95,9 @@ build_driver() {
   fi
 
   if [ -d drivers/"$2" ] ; then
-    ${TARGET}gcc -g -I drivers/"$2" drivers/libdrivers.c drivers/"$2"/*.c -o built_drivers/"$2".elf -T $1 $CFLAGS
+    ${GCC} -g -I drivers/"$2" drivers/libdrivers.c drivers/"$2"/*.c -o built_drivers/"$2".elf -T $1 $CFLAGS
   else
-    ${TARGET}gcc -g drivers/libdrivers.c drivers/"$2".c -o built_drivers/"$2".elf -T $1 $CFLAGS $SYSTEM_DEFINE
+    ${GCC} -g drivers/libdrivers.c drivers/"$2".c -o built_drivers/"$2".elf -T $1 $CFLAGS $SYSTEM_DEFINE
   fi &&
   # elf object file to binary, runnable code and data
   ${TARGET}objcopy -O binary built_drivers/"$2".elf built_drivers/"$2".bin --gap-fill 42 --pad-to $(( $( symbol built_drivers/"$2".elf pad_to_here ) )) &&
@@ -73,7 +105,8 @@ build_driver() {
   ${TARGET}objcopy -I binary -O elf64-littleaarch64 -B aarch64 built_drivers/"$2".bin built_drivers/"$2".o --rename-section .data=.data_driver,alloc,load,readonly,data,contents \
     --add-symbol '_binary_built_drivers_'$2'_data_pages='$(( $( symbol built_drivers/"$2".elf data_pages ) )) \
     --add-symbol '_binary_built_drivers_'$2'_code_pages='$(( $( symbol built_drivers/"$2".elf code_pages ) )) &&
-  ${TARGET}objdump -x --source --disassemble built_drivers/"$2".elf > built_drivers/"$2".dump
+  ${TARGET}objdump -x --source --disassemble built_drivers/"$2".elf > built_drivers/"$2".dump &&
+  echo Built $2
 }
 
 build_driver system_driver.ld.script "$SYSTEM_DRIVER" &&
@@ -103,7 +136,7 @@ echo '  integer_register start;'
 echo '  integer_register end;'
 echo '  integer_register code_pages;'
 echo '  integer_register data_pages;'
-echo '} drivers[] = {'
+echo '} const drivers[] = {'
 for driver in $SYSTEM_DRIVER $MEMORY_DRIVER $DRIVERS
 do
   if [ $driver != $SYSTEM_DRIVER ]; then echo "  ," ; fi
@@ -130,10 +163,14 @@ echo done: $STACK_SIZE
 
 echo Core size $CORE_SIZE
 
-${TARGET}gcc -o kernel8.elf -T ld.script -I built_drivers/ $KERNEL_ELEMENTS built_drivers/*.o $CFLAGS -DCORE_SIZE=\"$CORE_SIZE\" -DCORE_STACK_SIZE=$STACK_SIZE &&
+echo Compiling kernel
+${GCC} -o kernel8.elf -T ld.script -I built_drivers/ $KERNEL_ELEMENTS built_drivers/*.o $CFLAGS -DCORE_SIZE=\"$CORE_SIZE\" -DCORE_STACK_SIZE=$STACK_SIZE &&
+echo compiled kernel OK &&
 
 ${TARGET}objcopy kernel8.elf sdfat/kernel8.img -O binary &&
+echo ObjCopy OK &&
 ${TARGET}objdump -x --source --disassemble-all kernel8.elf > kernel8.dump &&
+echo ObjDump OK &&
 
 # Check for common mistakes in the build or code
 echo Checking that all variables are initialised in the kernel files &&
